AP-6372: FileStores use int[] (instead of byte[]) to denote loop path
AP-6372 (Parallel Chunk loop with large parallelization and nested
scopes might error with: "Too many nested loops: 128")

+++ org.knime.core/src/eclipse/org/knime/core/data/filestore/internal/WriteFileStoreHandler.java
@@ -270,7 +270,7 @@ public class WriteFileStoreHandler implements IWriteFileStoreHandler {
     }
 
     FileStore createFileStoreInternal(final String name,
-            final byte[] nestedLoopPath, final int iterationIndex) throws IOException {
+            final int[] nestedLoopPath, final int iterationIndex) throws IOException {
         assert Thread.holdsLock(this);
         CheckUtils.checkArgumentNotNull(name, "Argument must not be null.");
         if (name.startsWith(".")) {
